requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bayesian-optimization==1.4.2
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
chai-sacred==0.8.3
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
docopt==0.6.2
executing==1.2.0
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.49.1
gym==0.21.0
huggingface-hub==0.13.3
huggingface-sb3==2.2.4
hydra-core==1.3.2
hyperopt==0.2.7
idna==3.4
imitation==0.3.2
importlib-metadata==6.3.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.5
munch==2.5.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.27
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
ray==2.3.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seals==0.1.5
six==1.16.0
smmap==5.0.0
stable-baselines3==1.7.0
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.22.0
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
